tests/wpt/web-platform-tests/web-nfc/NDEFRecord_constructor.https.html

<!DOCTYPE html>
<title>NDEFRecord constructor</title>
<link rel="help" href="https://w3c.github.io/web-nfc/#dom-ndefrecord">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/nfc-helpers.js"></script>
<script>

  test(() => {
    assert_equals(NDEFRecord.length, 1);
    assert_throws(new TypeError, () => new NDEFRecord());
  }, 'NDEFRecord constructor without init dict');

  test(() => {
    assert_throws(new TypeError, () => new NDEFRecord(null),
        'The record has neither type nor data.');
  }, 'NDEFRecord constructor with null init dict');

  test(() => {
    const record = new NDEFRecord(createTextRecord(test_text_data));
    assert_equals(record.recordType, 'text', 'recordType');
    assert_equals(record.mediaType, 'text/plain', 'mediaType');
    const decoder = new TextDecoder();
    assert_equals(decoder.decode(record.data), test_text_data,
        'data has the same content with the original dictionary');
    assert_equals(record.text(), test_text_data,
        'text() has the same content with the original dictionary');
  }, 'NDEFRecord constructor with text record type');

  test(() => {
    const record = new NDEFRecord(createUrlRecord(test_url_data));
    assert_equals(record.recordType, 'url', 'recordType');
    assert_equals(record.mediaType, 'text/plain', 'mediaType');
    const decoder = new TextDecoder();
    assert_equals(decoder.decode(record.data), test_url_data,
        'data has the same content with the original dictionary');
    assert_equals(record.text(), test_url_data,
        'text() has the same content with the original dictionary');
  }, 'NDEFRecord constructor with url record type');

  test(() => {
    let buffer = new ArrayBuffer(4);
    let buffer_view = new Uint8Array(buffer);
    let original_data = new Uint8Array([1, 2, 3, 4]);
    buffer_view.set(original_data);
    const record = new NDEFRecord(createOpaqueRecord(buffer));
    assert_equals(record.recordType, 'opaque', 'recordType');
    assert_equals(record.mediaType, 'application/octet-stream', 'mediaType');
    assert_array_equals(new Uint8Array(record.data.buffer), original_data,
        'data has the same content with the original buffer');

    const data_1 = record.arrayBuffer();
    assert_true(data_1 instanceof ArrayBuffer);
    assert_not_equals(data_1, buffer, 'arrayBuffer() returns a new object');
    assert_array_equals(new Uint8Array(data_1), original_data,
        'arrayBuffer() has the same content with the original buffer');

    const data_2 = record.arrayBuffer();
    assert_true(data_2 instanceof ArrayBuffer);
    assert_not_equals(data_2, data_1,
        'arrayBuffer() again returns another new object');
    assert_array_equals(new Uint8Array(data_2), original_data,
        'arrayBuffer() has the same content with the original buffer');

    buffer_view.set([4, 3, 2, 1]);
    const data_3 = record.arrayBuffer();
    assert_true(data_3 instanceof ArrayBuffer);
    assert_array_equals(new Uint8Array(data_1), original_data,
        'Modifying the original buffer does not affect arrayBuffer() content');
    assert_array_equals(new Uint8Array(data_2), original_data,
        'Modifying the original buffer does not affect arrayBuffer() content');
    assert_array_equals(new Uint8Array(data_3), original_data,
        'Modifying the original buffer does not affect arrayBuffer() content');
  }, 'NDEFRecord constructor with opaque record type');

  test(() => {
    const record = new NDEFRecord(createJsonRecord(test_json_data));
    assert_equals(record.recordType, 'json', 'recordType');
    assert_equals(record.mediaType, 'application/json', 'mediaType');

    const data_1 = record.json();
    assert_true(typeof data_1 === 'object');
    assert_not_equals(data_1, test_json_data, 'json() returns a new object');
    assert_object_equals(data_1, test_json_data,
        'json() has the same content with the original dictionary');

    const data_2 = record.json();
    assert_true(typeof data_2 === 'object');
    assert_not_equals(data_2, data_1,
        'json() again returns another new object');
    assert_object_equals(data_2, test_json_data,
        'json() has the same content with the original dictionary');
  }, 'NDEFRecord constructor with JSON record type');

  test(() => {
    let buffer = new ArrayBuffer(4);
    let buffer_view = new Uint8Array(buffer);
    let original_data = new Uint8Array([1, 2, 3, 4]);
    buffer_view.set(original_data);
    const record = new NDEFRecord(createRecord('foo.eXamPle.coM:bAr*-', undefined, buffer));
    assert_equals(record.recordType, 'foo.example.com:bAr*-', 'recordType');
    assert_equals(record.mediaType, 'application/octet-stream', 'mediaType');

    const data_1 = record.arrayBuffer();
    assert_true(data_1 instanceof ArrayBuffer);
    assert_not_equals(data_1, buffer, 'arrayBuffer() returns a new object');
    assert_array_equals(new Uint8Array(data_1), original_data,
        'arrayBuffer() has the same content with the original buffer');

    const data_2 = record.arrayBuffer();
    assert_true(data_2 instanceof ArrayBuffer);
    assert_not_equals(data_2, data_1,
        'arrayBuffer() again returns another new object');
    assert_array_equals(new Uint8Array(data_2), original_data,
        'arrayBuffer() has the same content with the original buffer');

    buffer_view.set([4, 3, 2, 1]);
    const data_3 = record.arrayBuffer();
    assert_true(data_3 instanceof ArrayBuffer);
    assert_array_equals(new Uint8Array(data_1), original_data,
        'Modifying the original buffer does not affect arrayBuffer() content');
    assert_array_equals(new Uint8Array(data_2), original_data,
        'Modifying the original buffer does not affect arrayBuffer() content');
    assert_array_equals(new Uint8Array(data_3), original_data,
        'Modifying the original buffer does not affect arrayBuffer() content');
  }, 'NDEFRecord constructor with external record type');

  test(() => {
    assert_throws(new TypeError, () => new NDEFRecord(createRecord('EMptY')),
        'Unknown record type.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord('TeXt', '', test_text_data)),
        'Unknown record type.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord('uRL', '', test_url_data)),
        'Unknown record type.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord('jSoN', '', test_json_data)),
        'Unknown record type.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord('OpaQUE', '', test_buffer_data)),
        'Unknown record type.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord('sMart-PosTER', '', test_url_data)),
        'Unknown record type.');
  }, 'NDEFRecord constructor with record type string being treated as case sensitive');

  test(() => {
    assert_throws(new TypeError, () => new NDEFRecord(createRecord(
        ':xyz', '', test_buffer_data)), 'The domain should not be empty.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord(
        '[:xyz', '', test_buffer_data)), '"[" is not a valid FQDN.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord(
        'example.com:', '', test_buffer_data)), 'The type should not be empty.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord(
        'example.com:xyz~', '', test_buffer_data)), 'The type should not contain \'~\'.');
    assert_throws(new TypeError, () => new NDEFRecord(createRecord(
        'example.com:xyz/', '', test_buffer_data)), 'The type should not contain \'/\'.');
  }, 'NDEFRecord constructor with invalid external record type');

</script>
